+++ packages/vips/src/worker.ts
@@ -22,6 +22,7 @@ import {
 	batchResizeImage,
 	rotateImage,
 	hasTransparency,
+	getUltraHdrInfo,
 } from './index';
 
 /**
@@ -35,6 +36,7 @@ const api = {
 	batchResizeImage,
 	rotateImage,
 	hasTransparency,
+	getUltraHdrInfo,
 };
 
 expose( api );