+++ packages/core-data/src/utils/crdt-user-selections.ts
@@ -12,7 +12,11 @@ import { store as blockEditorStore } from '@wordpress/block-editor';
 import { CRDT_RECORD_MAP_KEY } from '../sync';
 import type { YPostRecord } from './crdt';
 import type { YBlock, YBlocks } from './crdt-blocks';
-import { getRootMap, richTextOffsetToHtmlIndex } from './crdt-utils';
+import {
+	asRichTextOffset,
+	getRootMap,
+	richTextOffsetToHtmlIndex,
+} from './crdt-utils';
 import type {
 	AbsoluteBlockIndexPath,
 	WPBlockSelection,
@@ -178,7 +182,10 @@ function getCursorPosition(
 
 	const relativePosition = Y.createRelativePositionFromTypeIndex(
 		currentYText,
-		richTextOffsetToHtmlIndex( currentYText.toString(), selection.offset )
+		richTextOffsetToHtmlIndex(
+			currentYText.toString(),
+			asRichTextOffset( selection.offset )
+		)
 	);
 
 	return {