+++ routes/taxonomies/actions/index.ts
@@ -1,3 +1,0 @@
-export { default as deleteTaxonomyAction } from './delete';
-export { default as editTaxonomyAction } from './edit';
-export { default as toggleActiveAction } from './toggle-active';